Changed the "quotes" rule from single-quotes to double-quotes.  Use ESLint's `--fix` option to automatically make this change

+++ style.js
@@ -155,7 +155,7 @@ module.exports = {
     // require single quotes for all strings
     quotes: [
       'error',
-      'single',
+      'double',
       'avoid-escape',
     ],
